refactor(api): split provider parts and document the stream contract
Normal imports replace the part-of facade; the ~1500 openai_common target was wrong.
