+++ lib/core/services/api/providers/openai_images.dart
@@ -1,7 +1,21 @@
-part of '../chat_api_service.dart';
+import 'dart:async';
+import 'dart:convert';
+import 'dart:io';
 
-bool _shouldUseOpenAIImagesApi(ProviderConfig config, String modelId) {
-  final upstreamModelId = _apiModelId(config, modelId).toLowerCase();
+import 'package:http/http.dart' as http;
+import 'package:http_parser/http_parser.dart';
+
+import '../../../models/token_usage.dart';
+import '../../../providers/settings_provider.dart';
+import '../../../utils/multimodal_input_utils.dart';
+import '../../../../utils/app_directories.dart';
+import '../../../../utils/sandbox_path_resolver.dart';
+import '../chat_api_helpers.dart';
+import '../stream/stream_chunk.dart';
+import '../stream/stream_chunk_emit.dart';
+
+bool shouldUseOpenAIImagesApi(ProviderConfig config, String modelId) {
+  final upstreamModelId = apiModelId(config, modelId).toLowerCase();
   return _supportsOpenAIImageGenerations(upstreamModelId);
 }
 
@@ -29,7 +43,7 @@ Uri _openAIImagesUrl(ProviderConfig config, String path) {
   return Uri.parse('$rawBase$path');
 }
 
-Stream<StreamChunk> _sendOpenAIImagesStream(
+Stream<StreamChunk> sendOpenAIImagesStream(
   http.Client client,
   ProviderConfig config,
   String modelId,
@@ -40,7 +54,7 @@ Stream<StreamChunk> _sendOpenAIImagesStream(
 }) async* {
   final input = await _openAIImagesInput(messages, userImagePaths);
   final outputMime = _openAIImagesOutputMime(config, modelId, extraBody);
-  final upstreamModelId = _apiModelId(config, modelId);
+  final upstreamModelId = apiModelId(config, modelId);
   if (input.imageRefs.isNotEmpty &&
       !_supportsOpenAIImageEdits(upstreamModelId)) {
     throw UnsupportedError(
@@ -83,7 +97,7 @@ Future<Map<String, dynamic>> _sendOpenAIImageGeneration(
   Map<String, dynamic>? extraBody,
 }) async {
   final body = <String, dynamic>{
-    'model': _apiModelId(config, modelId),
+    'model': apiModelId(config, modelId),
     'prompt': prompt,
   };
   _applyOpenAIImagesExtraBody(body, config, modelId, extraBody);
@@ -104,14 +118,14 @@ Future<Map<String, dynamic>> _sendOpenAIImageEdit(
   ProviderConfig config,
   String modelId,
   String prompt,
-  List<_ImageRef> imageRefs, {
+  List<ImageRef> imageRefs, {
   Map<String, String>? extraHeaders,
   Map<String, dynamic>? extraBody,
 }) async {
   final allRemote = imageRefs.every((ref) => ref.kind == 'url');
   if (allRemote) {
     final body = <String, dynamic>{
-      'model': _apiModelId(config, modelId),
+      'model': apiModelId(config, modelId),
       'prompt': prompt,
       'images': [
         for (final ref in imageRefs) {'image_url': ref.src},
@@ -143,7 +157,7 @@ Future<Map<String, dynamic>> _sendOpenAIImageEdit(
   request.headers.addAll(
     _openAIImagesMultipartHeaders(config, modelId, extraHeaders: extraHeaders),
   );
-  request.fields['model'] = _apiModelId(config, modelId);
+  request.fields['model'] = apiModelId(config, modelId);
   request.fields['prompt'] = prompt;
   final body = <String, dynamic>{};
   _applyOpenAIImagesExtraBody(body, config, modelId, extraBody);
@@ -183,7 +197,7 @@ Future<String> _lastOpenAIImagePrompt(
       if (prompt.isNotEmpty) return prompt;
       continue;
     }
-    final parsed = await _parseTextAndImages(
+    final parsed = await parseTextAndImages(
       (content ?? '').toString(),
       allowRemoteImages: true,
       allowLocalImages: true,
@@ -219,7 +233,7 @@ Future<_OpenAIImagesInput> _openAIImagesInput(
         prompt: prompt,
         imageRefs: [
           for (final mediaRef in internalMediaRefs)
-            if (isImageMime(_mimeForInternalMediaRef(mediaRef)))
+            if (isImageMime(mimeForInternalMediaRef(mediaRef)))
               _imageRefFromSource(mediaRef.uri, mime: mediaRef.mime),
         ],
       );
@@ -248,7 +262,7 @@ Future<_OpenAIImagesInput> _openAIImagesInput(
       }
     }
 
-    final parsed = await _parseTextAndImages(
+    final parsed = await parseTextAndImages(
       (content ?? '').toString(),
       allowRemoteImages: true,
       allowLocalImages: true,
@@ -271,7 +285,7 @@ Future<_OpenAIImagesInput> _openAIImagesInput(
   return _OpenAIImagesInput(prompt: prompt);
 }
 
-_ImageRef? _lastAssistantImageBefore(
+ImageRef? _lastAssistantImageBefore(
   List<Map<String, dynamic>> messages,
   int beforeIndex,
 ) {
@@ -283,9 +297,9 @@ _ImageRef? _lastAssistantImageBefore(
   return null;
 }
 
-List<_ImageRef> _extractOpenAIImageRefs(dynamic content) {
+List<ImageRef> _extractOpenAIImageRefs(dynamic content) {
   if (content is List) {
-    final refs = <_ImageRef>[];
+    final refs = <ImageRef>[];
     for (final part in content) {
       if (part is! Map) continue;
       final type = (part['type'] ?? '').toString();
@@ -300,8 +314,8 @@ List<_ImageRef> _extractOpenAIImageRefs(dynamic content) {
   }
 
   final raw = (content ?? '').toString();
-  if (raw.isEmpty) return const <_ImageRef>[];
-  final refs = <_ImageRef>[];
+  if (raw.isEmpty) return const <ImageRef>[];
+  final refs = <ImageRef>[];
   // Markdown images only. Custom attachment markers are not recognized;
   // attachments arrive via userImagePaths / multimodalInternalMediaPathsKey.
   final markdownImage = RegExp(r'!\[[^\]]*\]\(([^)]+)\)');
@@ -312,7 +326,7 @@ List<_ImageRef> _extractOpenAIImageRefs(dynamic content) {
   return refs;
 }
 
-void _addOpenAIStructuredImageRefs(List<_ImageRef> refs, dynamic value) {
+void _addOpenAIStructuredImageRefs(List<ImageRef> refs, dynamic value) {
   if (value == null) return;
   if (value is List) {
     for (final item in value) {
@@ -342,7 +356,7 @@ void _addOpenAIStructuredImageRefs(List<_ImageRef> refs, dynamic value) {
 }
 
 void _addOpenAIStructuredImageData(
-  List<_ImageRef> refs,
+  List<ImageRef> refs,
   dynamic data, {
   required bool isBase64,
   required String mime,
@@ -361,21 +375,21 @@ void _addOpenAIStructuredImageData(
   refs.add(_imageRefFromSource(source));
 }
 
-_ImageRef _imageRefFromSource(String source, {String? mime}) {
+ImageRef _imageRefFromSource(String source, {String? mime}) {
   if (source.startsWith('data:')) {
-    return _ImageRef('data', source, mime: mime);
+    return ImageRef('data', source, mime: mime);
   }
   if (source.startsWith('http://') || source.startsWith('https://')) {
-    return _ImageRef('url', source, mime: mime);
+    return ImageRef('url', source, mime: mime);
   }
-  return _ImageRef('path', source, mime: mime);
+  return ImageRef('path', source, mime: mime);
 }
 
-Future<http.MultipartFile?> _tryOpenAIImageMultipartFile(_ImageRef ref) async {
+Future<http.MultipartFile?> _tryOpenAIImageMultipartFile(ImageRef ref) async {
   if (ref.kind == 'data') {
     final mime = (ref.mime != null && ref.mime!.trim().isNotEmpty)
         ? ref.mime!.trim()
-        : _mimeFromDataUrl(ref.src);
+        : mimeFromDataUrl(ref.src);
     final commaIndex = ref.src.indexOf(',');
     final payload = commaIndex >= 0
         ? ref.src.substring(commaIndex + 1)
@@ -393,7 +407,7 @@ Future<http.MultipartFile?> _tryOpenAIImageMultipartFile(_ImageRef ref) async {
     if (!await file.exists()) return null;
     final mime = (ref.mime != null && ref.mime!.trim().isNotEmpty)
         ? ref.mime!.trim()
-        : _mimeFromPath(fixed);
+        : mimeFromPath(fixed);
     return http.MultipartFile.fromPath(
       'image[]',
       fixed,
@@ -421,11 +435,11 @@ Map<String, String> _openAIImagesJsonHeaders(
   String modelId, {
   Map<String, String>? extraHeaders,
 }) {
-  return _customHeaders(
+  return customHeaders(
     config,
     modelId,
     baseHeaders: <String, String>{
-      'Authorization': 'Bearer ${_apiKeyForRequest(config, modelId)}',
+      'Authorization': 'Bearer ${apiKeyForRequest(config, modelId)}',
       'Content-Type': 'application/json',
     },
     assistantHeaders: extraHeaders,
@@ -437,11 +451,11 @@ Map<String, String> _openAIImagesMultipartHeaders(
   String modelId, {
   Map<String, String>? extraHeaders,
 }) {
-  final headers = _customHeaders(
+  final headers = customHeaders(
     config,
     modelId,
     baseHeaders: <String, String>{
-      'Authorization': 'Bearer ${_apiKeyForRequest(config, modelId)}',
+      'Authorization': 'Bearer ${apiKeyForRequest(config, modelId)}',
     },
     assistantHeaders: extraHeaders,
   );
@@ -455,7 +469,7 @@ void _applyOpenAIImagesExtraBody(
   String modelId,
   Map<String, dynamic>? extraBody,
 ) {
-  final custom = _customBody(config, modelId, assistantBody: extraBody);
+  final custom = customBody(config, modelId, assistantBody: extraBody);
   if (custom.isNotEmpty) body.addAll(custom);
 }
 
@@ -544,5 +558,5 @@ class _OpenAIImagesInput {
   const _OpenAIImagesInput({required this.prompt, this.imageRefs = const []});
 
   final String prompt;
-  final List<_ImageRef> imageRefs;
+  final List<ImageRef> imageRefs;
 }
