+++ lib/core/services/api/stream/stream_chunk_decoder.dart
@@ -10,11 +10,11 @@ abstract class StreamChunkDecoder {
   /// Implementations should throw when the payload cannot be parsed.
   DecodeResult accept(SseEvent event);
 
-  /// Flush remaining state when the SSE connection closes normally.
+  /// Flush remaining open series when the SSE connection closes.
   ///
-  /// Must be idempotent with an explicit terminal event: if [accept] already
-  /// emitted [Finish], this returns an empty list. If the stream ended
-  /// without a terminal event, this emits a single [Finish].
+  /// Idempotent: a second call returns an empty list. Do not emit [Finish]
+  /// here — the provider emits [Finish] after [onClosed]. A second call
+  /// after an explicit terminal event must also be empty.
   List<StreamChunk> onClosed();
 }
